drivers: mali-valhall: Restart devfreq polling unconditionally on System resume
GPUCORE-32837: Restart devfreq polling unconditionally on System resume

Devfreq framework is used by Kbase for DVFS with a polling interval of
100 ms. So the GPU utilization is monitored every 100ms and the GPU
frequency is adjusted according to the utilization.

The devfreq polling is stopped on runtime suspend & system suspend and
is restarted on runtime resume & system resume. However the polling was
being restarted on system resume only if the work was submitted to the
GPU from system resume callback.
This conditional restart caused a problem for the scenario where GPU was
in runtime active state at the time of system suspend with no queued
jobs and the jobs were submitted immediately after system resume.
In that case the devfreq polling did not restart and GPU executed the
jobs at low frequency until the next power cycle.

This commit removes the condition for the restart of polling on system
resume.

BUG=b:207077408
TEST=emerge cherry kernel

Change-Id: I30844d42263dc3c6c934019dd9a3de3aaf741e23
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3520931
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Commit-Queue: Hsin-Yi Wang <hsinyi@chromium.org>